fix:更新

+++ Navigation/Navigation/Navigation/Private/Ground.h
@@ -10,29 +10,29 @@ public:
 	~Ground() {
 	};
 
-	virtual IGround * Init(int x, int y, int h, int f);
+	virtual IGround * Init(int x, int y, int h, short f);
 	virtual int GetX() const;
 	virtual int GetY() const;
 	virtual int GetHeight() const;
-	virtual int GetFlag() const;
+	virtual short GetFlag() const;
 	virtual IGround * SetX(int x);
 	virtual IGround * SetY(int y);
 	virtual IGround * SetHeight(int h);
 	virtual IGround * AddHeight(int h);
 	virtual IGround * SubHeight(int h);
-	virtual IGround * SetFlag(int f);
-	virtual IGround * AddFlag(int f);
-	virtual IGround * SubFlag(int f);
-	virtual bool HasFlag(int f);
+	virtual IGround * SetFlag(short f);
+	virtual IGround * AddFlag(short f);
+	virtual IGround * SubFlag(short f);
+	virtual bool HasFlag(short f);
 
 private:
 	int _x;
 	int _y;
 	int _h;
-	int _f;
+	short _f;
 };
 
-inline IGround * Ground::Init(int x, int y, int h, int f)
+inline IGround * Ground::Init(int x, int y, int h, short f)
 {
 	_x = x;
 	_y = y;
@@ -56,7 +56,7 @@ inline int Ground::GetHeight() const
 	return _h;
 }
 
-inline int Ground::GetFlag() const
+inline short Ground::GetFlag() const
 {
 	return _f;
 }
@@ -91,26 +91,25 @@ inline IGround * Ground::SubHeight(int h)
 	return this;
 }
 
-inline IGround * Ground::SetFlag(int f)
+inline IGround * Ground::SetFlag(short f)
 {
 	_f = f;
 	return this;
 }
 
-inline IGround * Ground::AddFlag(int f)
+inline IGround * Ground::AddFlag(short f)
 {
 	_f |= f;
 	return this;
 }
 
-inline IGround * Ground::SubFlag(int f)
+inline IGround * Ground::SubFlag(short f)
 {
 	_f &= ~f;
 	return this;
 }
 
-inline bool Ground::HasFlag(int f)
+inline bool Ground::HasFlag(short f)
 {
 	return _f & f;
 }
-
